requirements_win.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyasynchat==1.0.4
pyasyncore==1.0.4
pyftpdlib==2.0.1
pygame==2.6.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyserial==3.5
python-vlc==3.0.21203
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==80.9.0
simpleaudio==1.0.4
